refactor: extract `FinalizedHeadTracker` into a separate module
