+++ chain-signatures/chain-ethereum/src/lib.rs
@@ -7,6 +7,7 @@ mod client;
 mod config;
 mod event_parsing;
 mod execution_watcher;
+mod finalized_head;
 mod indexer;
 mod indexer_eth_direct_rpc;
 #[cfg(feature = "helios")]
